requirements.txt
functions-framework
Flask
gunicorn
google-cloud-bigquery
google-cloud-secret-manager
requests
google-cloud-storage
google-cloud-core
google-cloud-workflows>=1.0.0
google-auth>=2.0.0
google-cloud-dataform
